plugin for babeljs to convert i18n jsx elements to function calls.
Summary:
Our i18n system uses function calls in our JavaScript code to mark
strings for translation (or not).  In jsx we use <$_></$_> and
<$i18nDoNoTranslate></$i18nDoNoTranslate> to make strings for
translation (or not).  The plugin converts those tags to function
calls when using babeljs to transpile ES6+JSX code to ES5.

Test Plan: npm run test

Reviewers: aria

Subscribers: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D18143