Update plugin to babel 6
Summary:
The reason for updating the plugin is that we'd like try using flow
with webapp, but the current version of flow requires babel 6.

The fixtures were updated to include "use strict" at the start.  This
is b/c I wasn't able to figure out how to disable strict mode in babel 6
and this was not for trying. :(

Test Plan: - npm test

Reviewers: charlie, jared

Subscribers: colin

Differential Revision: https://phabricator.khanacademy.org/D31399